+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Features
 - Introduced function `requireResolved()`, which fails early if an environment variable reference could not be resolved ([#12](https://github.com/dedica-team/config-functions/pull/12))
+- Introduced function `urlEncode()`, which encodes a value for safe use in URLs ([#13](https://github.com/dedica-team/config-functions/pull/13))
 
 Development
 - Refactoring: Introduced `ConfigFunction` interface as a mechanism to simplify adding new functions ([#10](https://github.com/dedica-team/config-functions/pull/10))